wrote report APIs #5756

+++ question_service/src/main/java/org/khasanof/question_service/dto/question_report/QuestionReportDetailDTO.java
@@ -26,7 +26,7 @@ public class QuestionReportDetailDTO extends GenericDTO {
     private QuestionEntity question;
     private List<ReportEntity> reports;
     private Integer count;
-    private Long totalPointReports;
+    private Integer totalPointReports;
     private Instant updatedAt;
     private String updatedBy;
     private Instant createdAt;
